requirements.txt
django==3.2.8
djangorestframework==3.12.4
pillow==8.2.0
drf-nested-routers==0.93.4
djangorestframework-simplejwt==5.0.0
django-allauth==0.46.0
dj-rest-auth==2.1.11
django-enumchoicefield==2.0.0
django-cors-headers==3.7.0
python-dotenv==0.19.0
mysqlclient==2.0.3